requirements.txt
docstring-parser==0.15
Flask==2.3.3
Flask_Cors==4.0.0
SQLAlchemy==2.0.20
wonderwords==2.2.0
Flask-Restless-NG==3.0.0
Flask-SQLAlchemy==3.0.5
swagger-ui-py==23.8.23